add batch.score & move convenience to top batch function
